requirements.txt
djangorestframework # фреймворк джангорест
djangorestframework-simplejwt # JWT токены
psycopg2-binary # работа с psql
pillow # работа с картинками
python-decouple  # сокрытие данных
drf-yasg # документация для front
django-filter # фильтрация
